requirements.txt
pandas
numpy
scikit-learn
flask
streamlit
seaborn
matplotlib
dvc
pymysql
bentoml
python-dotenv
ipykernel
-e .